Login Button added
HandleClick now is its own function within the Login.jsx file, and if it finds "Enter" it doesn't add the value to the password but instead calls login, which no longer  has the enter key check in it. The form HTML tag was removed as it  was redirecting on its own, when we want to use the react-router redirect as we already are.